get back into 32f

+++ src/utils.cpp
@@ -14,7 +14,7 @@ cv::Mat helper :: sigmoid (const cv::Mat & X) {
 //     cv::exp(X, expX);
 
 //     cv::Mat sumExpX;
-//     cv::reduce(expX, sumExpX, 1, cv::REDUCE_SUM, CV_64FC1);
+//     cv::reduce(expX, sumExpX, 1, cv::REDUCE_SUM, CV_32FC1);
 
 //     cv::Mat softmaxResult;
 //     cv::divide(expX, repeat(sumExpX, 1, X.cols), softmaxResult);
@@ -25,13 +25,13 @@ cv::Mat helper :: sigmoid (const cv::Mat & X) {
 cv::Mat helper :: softmax(const cv::Mat &X) {
 
     cv::Mat maxVal;
-    cv::reduce(X, maxVal, 1, cv::REDUCE_MAX, CV_64FC1);
+    cv::reduce(X, maxVal, 1, cv::REDUCE_MAX, CV_32FC1);
 
     cv::Mat expX;
     cv::exp(X - repeat(maxVal, 1, X.cols), expX);
 
     cv::Mat sumExpX;
-    cv::reduce(expX, sumExpX, 1, cv::REDUCE_SUM, CV_64FC1);
+    cv::reduce(expX, sumExpX, 1, cv::REDUCE_SUM, CV_32FC1);
 
     cv::Mat softmaxResult = expX / repeat(sumExpX, 1, X.cols);
 
@@ -40,7 +40,7 @@ cv::Mat helper :: softmax(const cv::Mat &X) {
 
 
 cv::Mat helper :: relu (const cv::Mat & X) {
-    cv::Mat tmp(X.rows, X.cols, CV_64FC1);
+    cv::Mat tmp(X.rows, X.cols, CV_32FC1);
     for (int i=0; i<tmp.rows; i++){
         for(int j=0; j<tmp.cols; j++){
             if(tmp.at<float>(i, j) > 0.0)
@@ -60,7 +60,7 @@ cv::Mat helper :: relu (const cv::Mat & X) {
 cv::Mat helper :: sum (const cv::Mat & mat1, const cv::Mat & mat2) {
     assert(mat1.cols == mat2.cols);
 
-    cv::Mat tmp(mat1.rows, mat1.cols, CV_64FC1);
+    cv::Mat tmp(mat1.rows, mat1.cols, CV_32FC1);
 
     for (int i = 0; i < mat1.rows; i++) {
         for (int j = 0; j < mat1.cols; j++) {
@@ -79,7 +79,7 @@ cv::Mat helper :: sum (const cv::Mat & mat1, const cv::Mat & mat2) {
 cv::Mat helper :: sub (const cv::Mat & mat1, const cv::Mat & mat2) {
     assert(mat1.rows == mat2.rows);
 
-    cv::Mat tmp(mat1.rows, mat1.cols, CV_64FC1);
+    cv::Mat tmp(mat1.rows, mat1.cols, CV_32FC1);
 
     for (int i = 0; i < mat1.rows; i++) {
         for (int j = 0; j < mat1.cols; j++) {
@@ -123,7 +123,7 @@ double helper::categoricalCrossEntropy(const cv::Mat& y_train, const cv::Mat& y_
     double epsilon = 1e-15;
 
     // Create a matrix for one-hot encoding y_train
-    cv::Mat y_train_one_hot_labels(num_samples, num_classes, CV_64FC1, cv::Scalar(0.0));
+    cv::Mat y_train_one_hot_labels(num_samples, num_classes, CV_32FC1, cv::Scalar(0.0));
 
     for (int i = 0; i < num_samples; ++i) {
         int true_label = static_cast<int>(y_train.at<float>(i, 0));
